requirements.txt
fastmcp==2.10.6
aiohttp==3.12.14
pydantic==2.11.7
python-dotenv==1.1.1